+++ public/hero-image.png
